requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.127.0
google-generativeai==0.5.4
Flask==3.0.3
python-dotenv==1.0.1
rapidfuzz==3.9.3
tzdata==2024.1
plotly==5.24.1
kaleido==0.2.1
